requirements.txt
pandas
torch
torchvision
torch_geometric
ttach
tqdm
CairoSVG
matplotlib
opencv-python
networkx
numpy
ipython
rdkit
scikit-learn
lime
py3Dmol
streamlit
seaborn
plotly
wandb>=0.22.1,<0.23.0
python-dotenv>=0.9.9
loguru>=0.7.3,<0.8.0
shap>=0.48.0,<0.49.0
unidecode>=1.4.0,<2.0.0
biopython>=1.85,<2.0
captum>=0.8.0,<0.9.0
umap-learn>=0.5.9,<0.6.0
jupytext>=1.18.1,<2.0.0
pillow>=12.0.0,<13.0.0
kaleido>=1.2.0,<2.0.0
graphviz>=0.21
